LPS-174464 Avoid synchronization issues
I'm moving all the state to an AtomicReference so that it can never be read and
show an inconsistent content.

I don't want to use synchronized{}, nor ReadWriteLock for performance reasons and
because the possibility of finding state values in an inconsistent state, even if
this synchronization was missing, is almost insignificant.

(cherry picked from commit 2b0287fdc9f490108da2ad302ab5944dc79e83b9)