LPS-70835 Check if a page with same title already exists
After sending a page to trash, it is possible to create a page with
the same title.  As titles must be unique per group, we need to check
for this case before trying to sync using LARs.

Note that staging is not affected by this -- it is not possible to
edit live -- but site templates use LARs to propagate its changes to
sites.