LPS-174068 Avoiding mutations of the wrapped request
When extending HttpServletResponseWrapper some places in the code detect
it and unwrap the request to mutate it. By implementing all the methods
ourselves we guarantee no one can access the internal request.

This way we make sure the request cannot be mutated, at least directly.

I am not sure why this fixes the issue, because I don't really know what
the actual issue was, but it was caused because the request was being
mutated elsewhere.

I also made some tests by directly using the
PortalUtil.getOriginalServletRequest, and it was indeed being mutated
and things on the page randomly broke, which confirmed my theory that
there were unintended mutations. After these changes none of those
failures are happening.