LPS-168452 portal-bundle-blacklist-test: makes the test configuration multi-location
PROBLEM:

If a configuration is not given a location when it is created, it will be bound
to the bundle location of the first ManagedService service that is found.

Before the changes of LPS-162830, there was no ManagedService, so the
configuration would not be dynamically bound.

After the changes of LPS-162830, a ManagedService is registered for each
annotated configuration found in a bundle. This causes the configuration with no
location to be bound to the MangagedService's bundle in
com.liferay.portal.configuration.settings. This means that only listeners that
match that bundle will receive configuration updates.

The test waits for an update to the BundleBlacklist service whenever the
properties of the config are updated, but since the config is now bound to the
com.liferay.portal.configuration.settings bundle, the BundleBlacklist service
will no longer receive updates, and the test waits forever.

SOLUTION:

The solution to this is to pass in the "?" location when retrieving the
configuration. This avoids the dynamic binding and allows a listener in any
bundle to receive updates from this configuration.

See https://docs.osgi.org/specification/osgi.cmpn/7.0.0/service.cm.html#d0e11038